Styling
Edited Login.cshtml to move the twitter button to the center of the page
and added some style things to make the page visually appealing.

Edited the Content/Site.css to include styling for the homepage (login
page of the site)

Added image files to the content images folder for the Nebraska N and
the gradient background of the site.